The previous regexp hack didn't quite work; this one seems to
Finding the exact spot to tweak the qstring scanner config to accept or not
accept slash-quoted strings turned out to be harder than expected. The
voodoo in this commit works for simple cases; more testing is needed.

Also upped the priority of the comment scanner so that it has the highest of
the standard scanners. This makes sure that a /* */ comment is scanned as a
comment and not a slash-quoted string.