Refactored component id to component address.
ID will be the number with which a component will be referred to from
higher levels. Address is the number to which the component reacts.

This allows us to overload addresses by giving two components different
IDs but the same address, with one only handling XDATA signals and the
other one only XADDR signals. This provides a way around the limit
of 16 addresses.

Note that when actually implemented the term 'ID' may have morphed to
'Slot'. May.