Redo processor pipeline
The idea that a pipeline step can basically be boiled down to calling
a handler after transforming the node didn't work; in code generation
things often have to be done *before* the subordinate nodes are
transformed.

Inverted this paradigm; pipeline functions now switch on the node types
they're interested in, transform subordinate nodes and do their thing.
Node types that are not handled by the pipeline function should be
forwarded to process_tree().

In addition:
- Extended the code generator to handle more stuff
- Ripped a bunch of nonsense node types that were introduced to overcome
the above mentioned restriction
- Disabled the execute() pipeline function. Will probably come back