Pull up unary expressions with lower precedence
The expression @fh.error was being parsed as (@fh).error, which is
obviously wrong. Since '@' has a lower precedence than '.', the
binary should be done first, before the unary. So we push down the
binary and pull up the unary.

Also changed the 'as' keyword to a token; ideally 'as' would be an
operator, but right now it can't be because the rhs is a type which
we have to parse explicitly.