Improvements to the string datatype
Strings were implemented as Structs. This was nice because it allowed
us to map our language constructs onto C quite easily. However, this
makes optimized memory management hard.

So string is now it's own primitive type which maps to a new string
library which implements a light-weight shared pointer scheme for
strings.

Also implemented a destructor mechanism; types can define a destructor
operator tied to an intrinsic. Only string does that right now. A huge
TODO is to call the destructor when a variable goes out of scope.

Finally, all "our" C functions and variables that interact with
generated code should be prefixed with '$' to prevent clashes with
user code.